+++ crawler4j/src/main/java/edu/uci/ics/crawler4j/frontier/WebURLTupleBinding.java
@@ -38,7 +38,7 @@ public class WebURLTupleBinding extends TupleBinding<WebURL> {
         webURL.setDepth(input.readShort());
         webURL.setPriority(input.readByte());
         webURL.setAnchor(input.readString());
-        webURL.setFailedFetches(input.readInt());
+        webURL.setFailedFetches(input.readShort());
         return webURL;
     }
 
@@ -51,6 +51,6 @@ public class WebURLTupleBinding extends TupleBinding<WebURL> {
         output.writeShort(url.getDepth());
         output.writeByte(url.getPriority());
         output.writeString(url.getAnchor());
-        output.writeInt(url.getFailedFetches());
+        output.writeShort(url.getFailedFetches());
     }
 }
